Revert "progress bar should be in swift.go now i think"
This reverts commit 59413707f34323d5c9581221a7c560fe49b400b8.

I wasn't thinking clearly here, this moves it outside of where it should
be. Logically, I see what I was going for, but it doesn't make any sense
